requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
cryptography==43.0.0
etelemetry==0.3.1
filelock==3.15.4
fitz==0.0.1.dev2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.3
pillow==10.4.0
plum-dispatch==1.7.4
prov==2.0.1
pycparser==2.22
pydot==3.0.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.0
simplejson==3.19.3
six==1.16.0
spire==0.4.2
Spire.Pdf==10.7.1
SQLAlchemy==2.0.32
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
